+++ src/saveload/mod.rs
@@ -32,7 +32,7 @@ mod ser;
 mod tests;
 
 pub use self::de::{DeserializeComponents, FromDeserialize};
-pub use self::marker::{Marker, MarkerAllocator, U64Marker, U64MarkerAllocator};
+pub use self::marker::{MarkedBuilder, Marker, MarkerAllocator, U64Marker, U64MarkerAllocator};
 pub use self::ser::{IntoSerialize, SerializeComponents};
 
 /// A struct used for deserializing entity data.